changed polygons to use the distorted sprite algorithm
for the actual hardware, polygons are essentially identical to distorted sprites. the previous polygon algorithm didn't draw the curves that happen on the real hardware for concave shapes at all.  the current distorted sprite algorithm produces similar results to the hardware for distorted sprites and polygons, but that algorithm is not how the actual hardware draws.

the actual hardware draws using diagonal lines, which is why using half-transparent processing on distorted sprites and polygons is not recommended since the hardware overdraws to prevent gaps. thus, with half-transparent processing some pixels will be processed more than once, producing moire patterns in the drawn shapes.  

in order to have accurate goraud shading and endcode handling it may be necessary to correct the algorithm