Update einc immediately when changing registers it depends on in scsp.c
Make einc a pointer to the respective increment it has been set to so that
it updates when these change, and make sure to change the envelope
increments where necessary.

This introduces a slightly incompatibility with save states, and may result
in odd sound when switching formats until things settle.