Fixed window coordinates in software renderer.
That's kind of funny but we had a better function for
debugging window coordinates than the function actually
used for displaying windows...

Now both the debugging and "real" functions calls the
same code (with a thin wrapper for the debug part as
I didn't want to update everything...) to avoid the
problem in the future.

I also added something in vidsoft that looks like a hack
to fix window horizontal coordinates when in high res...
but I really think it's due to the way we handle hires.

This fix Cotton Boomerang selection screen.