Ports now have control over the OSD cores.
When compiled with the new YAB_PORT_OSD option, the emulation library
doesn't provide a OSDCoreList anymore and let ports choose their osd
core with the osdcoretype configuration in YabauseInit.
Of course, this means that trying to compile ports that don't handle
YAB_PORT_OSD will fail because of missing OSDCoreList symbol.