Fixed the previous fixes to manual mode (...)
Ten days without any change to those manual things was a bit too
long, so here's another change... this one tries to remove the
limit on frameskipping while not breaking anything else...

It "fixes" the frameskip for Choro Q and doesn't seems to break
anything in Panzer Dragoon (Zwei), King the Spirits, etc.

Please test with and without frameskipping, with both software
and opengl renderers.