Removed SetResolution / SetPriority functions.
This fix bugs in frameskip and maybe when changing video
core too. The idea is some games (and the bios) are not
setting priorities at each frame. That means only the
active video core got the correct priorities.

This is a somehow lazy change, most cores still have the
static SetResolution / SetPriority functions and still
use them internally.